Moved RepositoryFilter to filter package
