+++ eva-lib/src/main/java/uk/ac/ebi/eva/lib/repository/VariantEntityRepositoryCustom.java
@@ -22,7 +22,7 @@ import org.opencb.biodata.models.feature.Region;
 import org.springframework.data.domain.Pageable;
 
 import uk.ac.ebi.eva.commons.models.metadata.VariantEntity;
-import uk.ac.ebi.eva.lib.utils.RepositoryFilter;
+import uk.ac.ebi.eva.lib.filter.RepositoryFilter;
 
 import java.util.List;
